+++ run.sh
@@ -5,8 +5,7 @@ python3 deus.py \
     --transition_output_file transition_output.json \
     --loss_output_file loss_output.json \
     testinputs/shakemap.xml \
-    testinputs/exposure_sara.json \
+    testinputs/exposure_from_assetmaster.json \
     'SARA_v1.0' \
-    testinputs/fragility_sara.json \
-    testinputs/loss_sara.json
-#python3 deus.py --updated_exposure_output_file updated_exposure_output_file_ts.json testinputs/shakemap_tsunami.xml testinputs/exposure_supparsi.json 'SUPPARSI_2013.0' testinputs/fragility_supparsi.json 
+    testinputs/fragility_sara.json
+#python3 deus.py --updated_exposure_output_file updated_exposure_output_file_ts.json testinputs/shakemap_tsunami.xml testinputs/exposure_suppasri.json 'SARA_v1.0' testinputs/fragility_suppasri.json