+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2019-10-10:
+
+- Integrated the loss data files to be part of deus, so
+  the user does not have to give to the calls for deus anymore
+- Added aliases for intensity measurements
+
 # 2019-09-18
 
 - Added the changelog